Send the requestLoading: notification later.
According to the documentation in FBRequest.h the requestLoading:
notification is sent to the delegate right before the request gets sent
to the server. In the previous version the notification was sent before
the body of the request was built, and as there can be lengthy calls
during that process (such as UIImageJPEGRepresentation), there could
have been a significant delay between the requestLoading: notification
and the actual network request.